requirements.txt
fastapi==0.52.0
pydantic==1.4
pytest==5.4.1
python-dotenv==0.12.0
requests==2.23.0
starlette==0.13.2
tornado==5.1.1
uvicorn==0.11.3
python-jose==3.1.0

# Redis
redis==3.4.1
